requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
langchain==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
pinecone-client==4.1.1
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
langchain-pinecone==0.1.1
